# NTA/src/main/resources/apiSpec.yaml
openapi: 3.1.0
info:
  title: NTA Travel & Booking API
  version: 1.0.0
  description: 'API for managing Patients, Bookings, Support Persons, Service Providers, and Accommodations.'
  contact:
    name: Hamish Phillips
    email: hamish@phillipsmusictech.co.nz
servers:
  - url: 'http://localhost:8080/api'
paths:
...
  /support-persons:
    get:
      summary: List all support persons
      responses:
        '200':
          description: A list of support persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupportPerson'
        '404':
          description: Supporters Not Found
      operationId: getSupporters
      description: Retrieves all Patient Supporters
      tags:
        - support-persons
    post:
      summary: Create a new support person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportPerson'
      responses:
        '201':
          description: Support person created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportPerson'
        '409':
          description: Supporter Already Exists
      operationId: addSupportPerson
      description: Add a Support Person
      tags:
        - support-persons
  '/support-persons/{id}':
    get:
      summary: Get a support person by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Support person details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportPerson'
        '404':
          description: Supporter Not Found
      operationId: getSupporterById
      description: Retrieve a Support by thier ID.
      tags:
        - support-persons
    put:
      summary: Update a support person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportPerson'
      responses:
        '200':
          description: Updated support person
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportPerson'
        '404':
          description: Supporter Not Found
      operationId: updateSupporterById
      description: Updates a Supporter By thier ID.
      tags:
        - support-persons
    delete:
      summary: Delete a support person
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Support person deleted
        '404':
          description: Supporter Not Found
      operationId: deleteSupporterById
      description: Deletes a supporter by their ID.
      tags:
        - support-persons
...
components:
  schemas:
    Patient:
      type: object
      properties:
        nhi:
          type: string
          maxLength: 7
        ntaNumber:
          type: string
          maxLength: 7
        firstName:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        notes:
          type: array
          x-stoplight:
            id: a5b9uo2h5tn2b
          items:
            $ref: '#/components/schemas/Note'
            x-stoplight:
              id: 7118i73waf800
        supportProviders:
          x-stoplight:
            id: tu0kke7d8sy77
          type: array
          items:
            $ref: '#/components/schemas/PatientServiceProvider'
            x-stoplight:
              id: 0i20xl1qen82s
      required:
        - nhi
        - firstName
        - surname
    SupportPerson:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        coveredByNta:
          type: boolean
        patient:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: sv805lhoqkij2
      required:
        - id
        - firstName
        - surname
        - coveredByNta
    Booking:
      type: object
      properties:
        id:
          type: integer
        dateOfDeparture:
          type: string
          format: date
        dateOfReturn:
          type: string
          format: date
          nullable: true
        destination:
          type: string
          x-stoplight:
            id: 9ralr1tinlcdr
        bookingStatus:
          type: string
          maxLength: 15
        estimatedCost:
          type: number
          format: float
        estimatedCostForPatient:
          type: number
          x-stoplight:
            id: eu7asntsiivnn
          format: float
        bookingCreatedAt:
          type: string
          format: date-time
        serviceProvider:
          $ref: '#/components/schemas/ServiceProvider'
          x-stoplight:
            id: zwt73mcuom7en
        accommodation:
          $ref: '#/components/schemas/Accommodation'
          x-stoplight:
            id: omsd7xtb6by5i
        supportPersons:
          type: array
          items:
            $ref: '#/components/schemas/BookedSupportPerson'
        notes:
          x-stoplight:
            id: q1n5a1szwt3d5
          type: array
          items:
            $ref: '#/components/schemas/Note'
            x-stoplight:
              id: 0wfqry889bsg8
        patientNhi:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: cf6695mumxqcj
      required:
        - dateOfDeparture
        - bookingStatus
        - estimatedCost
        - bookingCreatedAt
        - serviceProvider
        - patientNhi
    ServiceProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 100
      required:
        - name
        - address
    Accommodation:
      type: object
      properties:
        address:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 50
      required:
        - address
        - name
    BookedSupportPerson:
      title: BookedSupportPerson
      x-stoplight:
        id: wjwp2ksall9pj
      type: object
      properties:
        supportPerson:
          $ref: '#/components/schemas/SupportPerson'
          x-stoplight:
            id: w5a3zm38w0tjk
    Note:
      title: Note
      x-stoplight:
        id: t1hhkqpm11vef
      type: object
      description: Notes Model
      properties:
        id:
          type: string
        message:
          type: string
          x-stoplight:
            id: bm45upjpx1c6g
        booking:
          $ref: '#/components/schemas/Booking'
          x-stoplight:
            id: k5x237crt9duh
        patient:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: 3qjg5h0ofe2ju
      required:
        - message
    PatientServiceProvider:
      title: PatientServiceProvider
      x-stoplight:
        id: ya45sz5s3ky2n
      type: object
      properties:
        id:
          type: string
          x-stoplight:
            id: 8tjjqoynmkgdq
        serviceProvider:
          $ref: '#/components/schemas/ServiceProvider'
          x-stoplight:
            id: gaxqo9yozbgo0
        frequency:
          type: string
          x-stoplight:
            id: pqjt85fm823dg
      required:
        - serviceProvider
        - frequency
...